feat!: uses single module for graphql impl
