+++ Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "pxid"
-version = "0.3.0"
+version = "0.4.0"
 edition = "2021"
 description = "Prefixed Globally Unique Identifier"
-repository = "https://github.com/EstebanBorai/pxid"
+repository = "https://github.com/whizzes/pxid"
 categories = ["web-programming"]
-homepage = "https://github.com/EstebanBorai/pxid"
+homepage = "https://github.com/whizzes/pxid"
 keywords = ["network", "id", "generator", "uid"]
 license = "MIT"
 authors = ["Esteban Borai <estebanborai@gmail.com>"]
